Added separate class DrawCanvas
